Avoid explicit casts

+++ cms/locale/locale.py
@@ -222,8 +222,7 @@ class Translation:
             n /= self.PREFIX_FACTOR
             if n < self.PREFIX_FACTOR:
                 f = copy.copy(self.locale.decimal_formats[None])
-                # We need int because floor returns a float in py2.
-                d = int(2 - math.floor(math.log10(n)))
+                d = 2 - math.floor(math.log10(n))
                 f.frac_prec = (d, d)
                 return (self.gettext(unit)
                         % babel.numbers.format_decimal(n, format=f,
